Added async result check as rule 1201
When using the result of an async operation (returning generic Task<>), rule 1200 fires indicating setup should only be used for overridable members. This is technically true because Task<> is abstract, however, the message is not helpful/actionable by a developer.

Added new rule, `Moq1201`, that is triggered when the setup method is for a Task and that Task includes the `.Result`. The error reflects the guidance from Moq, which is to use the `.ReturnsAsync` method for setup.