Cache reference assemblies between tests to improve performance (#71)
The testing framework does a bunch of heavy lifting to resolve
references for set of `ReferenceAssemblies`, including potentially
running the NuGet client to download packages. This class caches the
ReferenceAssemblies class (which is thread-safe), so that package
resolution only happens once for a given configuration.

The is the same approach taken by
https://github.com/dotnet/roslyn-analyzers/blob/4d5fd9da36d64d4c3370b8813122e226844fc6ed/src/Test.Utilities/AdditionalMetadataReferences.cs.

This cuts test time from ~40 seconds to ~20 seconds on my machine.