Add Code coverage report (#79)
Add code coverage report to pipeline. Fixes #42.

Design is basically:

1. Capture code coverage files in `test.runsettings` to avoid embedding
settings / logic in the pipeline YAML and make it possible to generate
coverage locally the same as CI
- Currently uses a simple assembly include limited to
`Moq.Analyzers.dll`. Ideally we could say "all project references".
Added a TODO for that, as well as to discuss with the code coverage team
3. Add `CleanCoverageReport` and `GenerateCoverageReport` targets in
test MSBuild slice to handle merging the reports across test projects
4. Use the MSBuild task for `dotnet-reportgenerator` to merge code
coverage reports into:
    - A markdown summary for GitHub
    - An HTML report with line-coverage info
    - A merged Cobertura XML report for tool processing
5. Update GitHub Actions YAML so that:
    - Markdown summary is added to the summary of the run
    - Other coverage reports are added as a build artifact

I didn't add any functionality to fail below a certain threshold, as
tools generally use a static threshold, or do odd things to smuggle
state like creating a branch in the repo or storing information in a
gist, so leaving that for later.