Remove Regex from analyzers (#81)
Replace using a `Regex` on string-ified type names with the proper
Roslyn API (usually Identifier.Name) and do some 'straightening' of the
code using early-returns and pattern matching to reduce the
triangle-ness.

This is effectively a rewrite of
`AsShouldBeUsedOnlyForInterfaceAnalyzer`. There's a smaller refactor for
`CallbackSignatureShouldMatchMockedMethod*` as that code needs to be
shared between the analyzer and the code fix provider (will be handled
later)

This refactor has a few benefits:

1. It addresses the bug mentioned in #58 (testing with a newer version
is coming in another PR to keep code reviewable)
2. It improves performance a good amount using very basic Visual Studio
tracing (e.g. ~20 fewer string allocations and saves 30% [20 ms] on a
single call to Analyzer.Analyze())
3. It plumbs the `CancellationToken` into more places to be a good IDE
citizen

This PR also starts to break apart the monolithic `Helpers` class into
smaller components. In order to support the `.TryXXX` pattern I needed
to bring in a polyfill for `[NotNullWhen]` for nullability analysis. I'm
using PolySharp because, in addition to being the most popular polyfill,
it uses a source generator to add polyfills to the current assembly,
which avoid assembly loading overhead and the hassle of adding
dependencies to the analyzer.