Remove period from analyzer messages and inline in the analyzer classes (#86)
Fix RS1032 by removing the period from the end of the analyzer messages.
I also inlined the constants into the appropriate analyzer class instead
of in a shared constants class. This makes future refactoring of the
tests easier, as now the analyzer's ID is part of the analyzer itself.

I also filed #85 to track potential changes to our message and
categories.