Remove Microsoft.CodeAnalysis.CSharp.Analyzer.Testing (#102)
Remove `Microsoft.CodeAnalysis.CSharp.Analyzer.Testing` because it is
duplicative with `Microsoft.CodeAnalysis.CSharp.CodeFix.Testing`. This
is common when building analyzers that contain code fixes. From a
testing perspective, an analyzer test is just a code fix test with an
"empty" or "null pattern" code fixer.

This also unblocks #97, as the two packages otherwise need upgrading
together. However, cleaning up our dependencies is preferred to creating
a group in dependabot.