Bug/fix issue 143 (#221)
Add improved detections for overridable members in Moq1200. Rule 1201
recommends the user not use `.Setup` for async methods; however, in
cases where that is suppressed, Moq1200 was incorrectly indicating the
`.Result` property was not overridable. Moq1200 now shares the same
logic as 1201 and will not perform analysis when the condition is
observed.

## Changes

- Move detection for Async return to common location
- Update Moq1200 with detection logic from Moq1201
- Add test case from #143 to catch issue

Resolves #143