Move "upload binlogs" CI step earlier in pipeline and run even if failed (#273)
Currently, the "upload binlogs" step is too far down in the CI pipeline
(and also doesn't run if the build fails). Move it up a bit so that the
order of log upload mirrors the order of actions and add `success() ||
failure()` so they are uploaded even during build failures.