Move minimum support to sdk 8 (#295)
For v1.0 release, Moq.Analyzers will move to support the latest stable
.NET LTS SDK

Changes:

- Updated `Microsoft.CodeAnalysis.CSharp` and
`Microsoft.CodeAnalysis.CSharp.Workspaces` updated to 4.8. This is the
earliest .NET 8 SDK we can support. See
https://github.com/dotnet/roslyn/blob/main/docs/wiki/NuGet-packages.md?plain=1#L49
- Removed unnecessary `CancellationToken` parameters in benchmark helper
methods as a result of the package updates
- Updated SDK to support roll-forward behavior. This is to work around
having to keep up to patches on workstations (or worse, getting auto
patched because of a CVE and having things fail)