Add AnalyzerUtilities to NuGet package (#326)
This fixes a regression caused by #290.

`Microsoft.CodeAnalysis.AnalyzerUtilities` is not one of the assemblies
provided by the host (i.e. compiler), so we need to include it in our
own package.

I'm not sure how to write an automated test for this. Our current tests
don't use the NuGet package. Doing something like `dotnet format` in
theory should work, but when I tried that it passed even without this
change.

This does result in a new package contents baseline, so we can chalk
this one up to reviewer error and revisit if we find a better solution.